Patch docManager.open to better open in new tab
