+++ packages/docmanager-extension/package.json
@@ -38,6 +38,9 @@
   "dependencies": {
     "@jupyterlab/application": "^3.0.0-rc.12",
     "@jupyterlab/coreutils": "^5.0.0-rc.12",
+    "@jupyterlab/docmanager": "^3.0.0-rc.12",
+    "@jupyterlab/docregistry": "^3.0.0-rc.12",
+    "@jupyterlab/services": "^6.0.0-rc.12",
     "@lumino/algorithm": "^1.3.3"
   },
   "devDependencies": {
